refactor: prisma instace and repository structure folder
